picture-service

+++ taw-picture-service/src/main/java/com/taw/picture/controller/FileUploadController.java
@@ -1,4 +1,4 @@
-package com.taw.scene.controller;
+package com.taw.picture.controller;
 
 import java.io.File;
 import java.io.FileOutputStream;
@@ -10,6 +10,8 @@ import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 import org.apache.commons.io.IOUtils;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Controller;
@@ -22,16 +24,25 @@ import com.hawk.pub.web.HttpRequestHandler;
 import com.hawk.pub.web.HttpResponseHandler;
 import com.hawk.pub.web.SuccessResponse;
 import com.hawk.utility.check.CheckTools;
+import com.taw.picture.configure.PictureServiceConfigure;
+import com.taw.picture.service.PictureService;
+import com.taw.picture.service.PictureService.ComputeResult;
 import com.taw.pub.scene.request.UploadLengthParam;
 import com.taw.pub.scene.response.UploadLengthResponse;
-import com.taw.scene.configure.SceneServiceConfigure;
 
 @Controller
 public class FileUploadController {
 	
+	private Logger logger = LoggerFactory.getLogger(getClass());
+	
+	
+	
 	@Autowired
-	@Qualifier("sceneServiceConfigure")
-	private SceneServiceConfigure sceneServiceConfigure;
+	@Qualifier("pictureServiceConfigure")
+	private PictureServiceConfigure pictureServiceConfigure;
+	
+	@Autowired
+	private PictureService pictureService;
 	
 	
 
@@ -41,7 +52,7 @@ public class FileUploadController {
 		 * yyyyMMddHHmmss + 32位uuid
 		 */
 		String uuid = request.getParameter("uuid");
-		ComputeResult computeResult = computeDir(uuid);
+		ComputeResult computeResult = pictureService.computeDir(uuid);
 		String dir = computeResult.getDir();
 		String suffix = computeResult.getSuffix();	
 		/**
@@ -66,23 +77,59 @@ public class FileUploadController {
 		}
 		long byteArraySize = new Long(byteArraySizeStr);
 		
+		/**
+		 * 本次上传内容 在文件的起始位置
+		 */
+		String offsetStr = request.getParameter("offset");
+		if(StringUtils.isEmpty(byteArraySizeStr)){
+			throw new RuntimeException("offset is empty");
+		}
+		long offset = new Long(offsetStr);
+		if (offset > srcFileSize)
+			throw new RuntimeException("offset is wrong");
+		
 		
 		InputStream in = request.getInputStream();
 		mkdir(dir);
 		String filePath = dir + File.separator + uuid ;
 		File file = new File(filePath+".tmp");
+		
+		long maxSize = pictureServiceConfigure.getAllowFileSize();
+		if (srcFileSize > maxSize || file.length() >maxSize  )
+			throw new RuntimeException("The upload file size shouln't be over " + maxSize);
+		
 		FileOutputStream fos = null;
 		try{
-			fos = new FileOutputStream(file, true);
-			IOUtils.copyLarge(in, fos);
-		}finally{		
-			fos.close();
+			if (file.exists() && file.length() >= srcFileSize){
+				logger.info("Success to upload file = {}",file.getName());
+			}else{
+				fos = new FileOutputStream(file, true);
+				IOUtils.copyLarge(in, fos);
+			}
+		}finally{	
+			try {
+				if (fos != null)
+					fos.close();
+			} catch (Exception e) {
+				logger.error("failed to close file");
+			}
 		}
 		
+		
+		
 		if (file.length() >= srcFileSize){
+			logger.info("Success to upload file = {}",file.getName());
 			boolean  flag = file.renameTo(new File(filePath)); 
+			
+			/**
+			 * 创建一个success 后缀的 flag 文件，仅仅标识文件已经上传完成。
+			 */
+			File successFile = new File(filePath + ".success");
+			successFile.createNewFile();
 			if (!flag)
 				throw new RuntimeException("Failed to rename file");
+			else
+				logger.info("Success to rename file = {}",file.getName());
 		}
 		
 		HttpResponseHandler.handle(response, SuccessResponse.SUCCESS_RESPONSE);
@@ -94,7 +141,7 @@ public class FileUploadController {
 		CheckTools.check(param);
 
 		String uuid = param.getUuid();
-		String filePath = computeDir(uuid).getDir()+File.separator + uuid;
+		String filePath = pictureService.computeDir(uuid).getDir()+File.separator + uuid;
 		File file = new File(filePath+".tmp");
 		long size = 0;
 		if (file.exists()){
@@ -126,60 +173,11 @@ public class FileUploadController {
 	
 	
 	
-	private static class ComputeResult{
-		public String getSuffix() {
-			return suffix;
-		}
-		public void setSuffix(String suffix) {
-			this.suffix = suffix;
-		}
-		public String getDir() {
-			return dir;
-		}
-		public void setDir(String dir) {
-			this.dir = dir;
-		}
-		private String suffix;
-		private String dir;
-	}
 	
-	private  ComputeResult computeDir(String uuid){
-		if (StringUtils.isEmpty(uuid))
-			throw new RuntimeException("uuid is null");
-		uuid = uuid.trim();		
-		if (uuid.length() <48)
-			throw new RuntimeException("invalid uuid , uuid=yyyyMMddHHmmss+32uuid + .suffix");
-		
-		String[] strArray = uuid.split("\\.");
-		
-		if (strArray.length != 2)
-			throw new RuntimeException("invalid uuid , uuid=yyyyMMddHHmmss+32uuid + .suffix");
-		
-		ComputeResult computeResult  = new ComputeResult();
-		computeResult.setSuffix(strArray[1]);
-		uuid = strArray[0];
-		int length = uuid.length();
-		strArray = new String[]{uuid.substring(length-8,length-6),
-				uuid.substring(length-6,length-4),
-				uuid.substring(length-4,length-2),
-				uuid.substring(length-2,length)};
-		
-		String path = sceneServiceConfigure.getUploadDir();
-		for (String str : strArray){
-			path = path + File.separator + computeDirPart(str);
-		}
-		
-		computeResult.setDir(path);
-		return  computeResult;
-		
-	}
 	
-	private static  String computeDirPart(String str){
-		str = new Integer(Integer.parseInt(str, 16) % 16).toString();
-		if (str.length() == 1)
-			str = "0"+str;
-		return str;
-	}
+	
+	
+
